+++ view/index.js
@@ -363,24 +363,22 @@ function getEmptyGameTree() {
 function loadSettings() {
     $$('head link.userstyle').set('href', setting.stylesPath)
 
-    if (setting.get('view.fuzzy_stone_placement'))
-        $('goban').addClass('fuzzy')
-    if (setting.get('view.show_coordinates'))
-        $('goban').addClass('coordinates')
-    if (setting.get('view.show_next_moves'))
-        $('goban').addClass('variations')
-    if (setting.get('view.show_siblings'))
-        $('goban').addClass('siblings')
+    $('goban').toggleClass('fuzzy', setting.get('view.fuzzy_stone_placement'))
+    $('goban').toggleClass('animation', setting.get('view.animated_stone_placement'))
+    $('goban').toggleClass('coordinates', setting.get('view.show_coordinates'))
+    $('goban').toggleClass('variations', setting.get('view.show_next_moves'))
+    $('goban').toggleClass('siblings', setting.get('view.show_siblings'))
+
     if (setting.get('view.show_leftsidebar')) {
         document.body.addClass('leftsidebar')
         setLeftSidebarWidth(setting.get('view.leftsidebar_width'))
     }
+
     if (setting.get('view.show_graph') || setting.get('view.show_comments')) {
         document.body.addClass('sidebar')
         setSidebarArrangement(setting.get('view.show_graph'), setting.get('view.show_comments'))
+        setSidebarWidth(setting.get('view.sidebar_width'))
     }
-
-    setSidebarWidth(setting.get('view.sidebar_width'))
 }
 
 function prepareEditTools() {
@@ -901,8 +899,11 @@ function makeMove(vertex, sendCommand) {
         var li = $$('#goban .pos_' + vertex[0] + '-' + vertex[1])
         var direction = Math.floor(Math.random() * 9)
 
+        li.addClass('animate')
         for (var i = 0; i < 9; i++) li.removeClass('shift_' + i)
         li.addClass('shift_' + direction)
+        setTimeout(function() { li.removeClass('animate') }, 200)
+
         readjustShifts(vertex)
     }
 
@@ -1594,6 +1595,7 @@ function commitPreferences() {
 
     remote.getCurrentWindow().webContents.setAudioMuted(!setting.get('sound.enable'))
     setFuzzyStonePlacement(setting.get('view.fuzzy_stone_placement'))
+    setAnimatedStonePlacement(setting.get('view.animated_stone_placement'))
 
     // Save engines